Move logic that could be shared across each {ResourceName}DetailsViewModelImpl to ResourceDetailsViewModelImpl.kt and {ResourceName}NamesViewModelImpl to ResourceNamesViewModelImpl.kt.
