+++ app/src/main/java/com/davidread/starwarsdatabase/viewmodel/PersonDetailsViewModelImpl.kt
@@ -42,37 +42,7 @@ class PersonDetailsViewModelImpl @Inject constructor(
     private val filmsRemoteDataSource: FilmsRemoteDataSource,
     private val starshipsRemoteDataSource: StarshipsRemoteDataSource,
     private val vehiclesRemoteDataSource: VehiclesRemoteDataSource
-) : ResourceDetailsViewModel, ViewModel() {
-
-    /**
-     * Emits a [List] of [ResourceDetailListItem]s that should be shown on the UI.
-     */
-    override val resourceDetailsLiveData: MutableLiveData<List<ResourceDetailListItem>> =
-        MutableLiveData()
-
-    /**
-     * Emits whether a loading state should be shown on the UI.
-     */
-    override val showLoadingLiveData: MutableLiveData<Boolean> = MutableLiveData(false)
-
-    /**
-     * Emits whether an error state should be shown on the UI.
-     */
-    override val showErrorLiveData: MutableLiveData<Boolean> = MutableLiveData(false)
-
-    /**
-     * Container for managing resources used by `Disposable`s or their subclasses.
-     */
-    private val disposable: CompositeDisposable = CompositeDisposable()
-
-    /**
-     * Called when this `ViewModel` is no longer used and will be destroyed. Clears any
-     * subscriptions held by [disposable] to prevent this `ViewModel` from leaking.
-     */
-    override fun onCleared() {
-        disposable.clear()
-        super.onCleared()
-    }
+) : ResourceDetailsViewModelImpl() {
 
     /**
      * Sets up a subscription for getting the details of a single person from SWAPI to show in the
