+++ app/src/main/java/com/davidread/starwarsdatabase/viewmodel/PersonNamesViewModelImpl.kt
@@ -2,14 +2,11 @@ package com.davidread.starwarsdatabase.viewmodel
 
 import android.util.Log
 import androidx.annotation.IntRange
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import com.davidread.starwarsdatabase.datasource.PeopleRemoteDataSource
 import com.davidread.starwarsdatabase.model.view.ResourceNameListItem
 import com.davidread.starwarsdatabase.util.extractIDFromURL
 import com.davidread.starwarsdatabase.util.extractPageFromURL
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.schedulers.Schedulers
 import javax.inject.Inject
 
@@ -20,36 +17,7 @@ import javax.inject.Inject
  * fetching people data from SWAPI.
  */
 class PersonNamesViewModelImpl @Inject constructor(private val peopleRemoteDataSource: PeopleRemoteDataSource) :
-    ResourceNamesViewModel, ViewModel() {
-
-    /**
-     * Emits a [List] of [ResourceNameListItem]s that should be shown on the UI.
-     */
-    override val resourceNamesLiveData: MutableLiveData<List<ResourceNameListItem>> =
-        MutableLiveData()
-
-    /**
-     * Whether all [ResourceNameListItem]s have been fetched from SWAPI.
-     */
-    override val isAllResourceNamesRequestedLiveData: MutableLiveData<Boolean> =
-        MutableLiveData(false)
-
-    /**
-     * Next page of person names to fetch from SWAPI.
-     */
-    @IntRange(from = 1)
-    override var nextPage: Int = 1
-
-    /**
-     * [MutableList] of [ResourceNameListItem]s to be stored/modified here. Emitted via its `LiveData`
-     * each time it is updated.
-     */
-    private val personNames: MutableList<ResourceNameListItem> = mutableListOf()
-
-    /**
-     * Container for managing resources used by `Disposable`s or their subclasses.
-     */
-    private val disposable: CompositeDisposable = CompositeDisposable()
+    ResourceNamesViewModelImpl() {
 
     /**
      * Called when this `ViewModel` is initially created. It sets up the initial subscription for
@@ -57,15 +25,6 @@ class PersonNamesViewModelImpl @Inject constructor(private val peopleRemoteDataS
      */
     init {
         getResourceNames(nextPage)
-    }
-
-    /**
-     * Called when this `ViewModel` is no longer used and will be destroyed. Clears any
-     * subscriptions held by [disposable] to prevent this `ViewModel` from leaking.
-     */
-    override fun onCleared() {
-        disposable.clear()
-        super.onCleared()
     }
 
     /**
@@ -79,11 +38,11 @@ class PersonNamesViewModelImpl @Inject constructor(private val peopleRemoteDataS
             .observeOn(AndroidSchedulers.mainThread())
             .subscribeOn(Schedulers.io())
             .doOnSubscribe {
-                personNames.apply {
+                resourceNames.apply {
                     remove(ResourceNameListItem.Error)
                     add(ResourceNameListItem.Loading)
                 }
-                resourceNamesLiveData.postValue(personNames)
+                resourceNamesLiveData.postValue(resourceNames)
             }
             .subscribe(
                 { pageResponse ->
@@ -93,11 +52,11 @@ class PersonNamesViewModelImpl @Inject constructor(private val peopleRemoteDataS
                             name = personResponse.name
                         )
                     }
-                    personNames.apply {
+                    resourceNames.apply {
                         remove(ResourceNameListItem.Loading)
                         addAll(newPersonNames)
                     }
-                    resourceNamesLiveData.postValue(personNames)
+                    resourceNamesLiveData.postValue(resourceNames)
                     if (pageResponse.next == null) {
                         isAllResourceNamesRequestedLiveData.postValue(true)
                     }
@@ -111,11 +70,11 @@ class PersonNamesViewModelImpl @Inject constructor(private val peopleRemoteDataS
                     }
                 },
                 { throwable ->
-                    personNames.apply {
+                    resourceNames.apply {
                         remove(ResourceNameListItem.Loading)
                         add(ResourceNameListItem.Error)
                     }
-                    resourceNamesLiveData.postValue(personNames)
+                    resourceNamesLiveData.postValue(resourceNames)
                     Log.e(TAG, throwable.toString())
                 }
             )
